[REFACTOR] Add landing page for courses
The navigation tabs require the opening of a page whenever a tab is
clicked. Since the tab is not part of the navigation tree, the
navigation window will decent the navigation tree in a depth first
manner until it reaches the first page.

To control which page is opened when navigating to a tag, we just add an
index page at the first page the depth first search finds.